Hangman with Mıps
First, a random word is taken from the word box.
The user tries to guess the word after entering their name.
You have the right to enter wrong letters a total of 7times.
The man hangs on the 7th wrong and it's game over.
You have the right to guess the whole word at each
stage.
In our project, we designed the HANGMAN game using the
MIPS code in the mars emulator.